refactor(ui): extract team-sync/events/init-team-sync-events.ts
Move initTeamSyncEvents and its three button handlers (Create invite,
Join team, Sync now) into a dedicated events module. The render
pipeline no longer mixes DOM listener wiring with render logic.
team-sync.ts re-exports initTeamSyncEvents so the index.ts call site
stays on the same import path.
